EDHOC-05 with PRK_ID X5T or X5CHAIN

+++ os/lib/cbor/cbor.h
@@ -49,7 +49,8 @@ int cbor_put_text(uint8_t **buffer, char *text, uint8_t text_len);
 
 int cbor_put_array(uint8_t **buffer, uint8_t elements);
 
-int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, uint8_t bytes_len);
+//int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, uint8_t bytes_len);
+int cbor_put_bytes(uint8_t **buffer, uint8_t *bytes, size_t bytes_len);
 
 int cbor_put_map(uint8_t **buffer, uint8_t elements);
