+++ os/net/security/cose/cose-log.h
@@ -74,27 +74,27 @@ void cose_print_char(uint8_t *buff, size_t len);
 #define LOG_DBG_COSE_STR(data, len)  LOG_COSE_STR(LOG_LEVEL_DBG, data, len)
 
 static inline void
-cose_print_buff_8_dbg(uint8_t *buf, uint8_t len)
+cose_print_buff_8_dbg(uint8_t *buf, size_t len)
 {
   LOG_DBG_COSE_BUFF(buf, len);
 }
 static inline void
-cose_print_buff_8_info(uint8_t *buf, uint8_t len)
+cose_print_buff_8_info(uint8_t *buf, size_t len)
 {
   LOG_INFO_COSE_BUFF(buf, len);
 }
 static inline void
-cose_print_buff_8_err(uint8_t *buf, uint8_t len)
+cose_print_buff_8_err(uint8_t *buf, size_t len)
 {
   LOG_ERR_COSE_BUFF(buf, len);
 }
 static inline void
-cose_print_char_8_info(uint8_t *buf, uint8_t len)
+cose_print_char_8_info(uint8_t *buf, size_t len)
 {
   LOG_INFO_COSE_STR(buf, len);
 }
 static inline void
-cose_print_char_8_dbg(uint8_t *buf, uint8_t len)
+cose_print_char_8_dbg(uint8_t *buf, size_t len)
 {
   LOG_DBG_COSE_STR(buf, len);
 }
